Move role context
