+++ packages/react-ai/src/components/AIConversation/context/MessagesContext.tsx
@@ -6,6 +6,11 @@ type MessagesContextProps = ConversationMessage[] | undefined;
 export const MessagesContext =
   React.createContext<MessagesContextProps>(undefined);
 
+// role of the user sending the message, assistant or user
+export const RoleContext = React.createContext<
+  ConversationMessage['role'] | undefined
+>(undefined);
+
 export const MessagesProvider = ({
   children,
   messages,
