+++ packages/react-ai/src/components/AIConversation/context/index.ts
@@ -1,7 +1,7 @@
 import { ActionsContext } from './ActionsContext';
 import { AvatarsContext } from './AvatarsContext';
 import { InputContext } from './InputContext';
-import { MessagesContext } from './MessagesContext';
+import { MessagesContext, RoleContext } from './MessagesContext';
 import { SuggestedPromptsContext } from './SuggestedPromptsContext';
 import { MessageVariantContext } from './MessageVariantContext';
 
@@ -10,6 +10,7 @@ export {
   AvatarsContext,
   InputContext,
   MessagesContext,
+  RoleContext,
   SuggestedPromptsContext,
   MessageVariantContext,
 };
